Clean up unused code
